Refactor database queries to models file
