+++ safe_locking_service/locking_events/models.py
@@ -1,6 +1,12 @@
-from django.db import models
+from decimal import Decimal
+from typing import Dict, List, Optional, TypedDict
+
+from django.db import connection, models
+from django.db.backends.utils import CursorWrapper
 from django.db.models import Index, Q
 
+from eth_typing import ChecksumAddress
+from hexbytes import HexBytes
 from web3.types import EventData
 
 from gnosis.eth.django.models import (
@@ -9,6 +15,27 @@ from gnosis.eth.django.models import (
     Uint32Field,
     Uint96Field,
 )
+
+
+class LeaderBoardRow(TypedDict):
+    position: int
+    holder: ChecksumAddress
+    lockedAmount: Decimal
+    unlockedAmount: Decimal
+    withdrawnAmount: Decimal
+
+
+def fetch_all_from_cursor(cursor: CursorWrapper) -> List[LeaderBoardRow]:
+    """
+    Return all rows from a cursor as a dict.
+    Assume the column names are unique.
+
+    :param cursor:
+    :return: all rows from a cursor as a dict.
+    """
+    columns = [col[0] for col in cursor.description]
+
+    return [dict(zip(columns, row)) for row in cursor.fetchall()]
 
 
 class EthereumTxManager(models.Manager):
@@ -195,3 +222,89 @@ class StatusEventsIndexer(models.Model):
 
     def __str__(self):
         return f"EventIndexer: address={self.contract} deployed_block={self.deployed_block} last_indexed_block={self.last_indexed_block} "
+
+
+class LeaderBoard:
+    """
+    Class to group all the database queries related with leaderboard
+    """
+
+    @classmethod
+    def _get_leader_board_query(cls) -> str:
+        """
+        Get raw leaderboard SQL query
+
+        :return:
+        """
+        union_lock_unlock_withdraw_events = """
+                        (SELECT "locking_events_lockevent"."holder" AS "holder",
+                               "locking_events_lockevent"."amount" AS "amount",
+                                0 AS "event_type"
+                        FROM "locking_events_lockevent"
+                        ) UNION ALL (
+                        SELECT "locking_events_unlockevent"."holder" AS "holder",
+                               "locking_events_unlockevent"."amount" AS "amount",
+                                1 AS "event_type"
+                        FROM "locking_events_unlockevent")
+                        UNION ALL
+                        (SELECT "locking_events_withdrawnevent"."holder" AS "holder",
+                                "locking_events_withdrawnevent"."amount" AS "amount",
+                                2 AS "event_type"
+                        FROM "locking_events_withdrawnevent")
+                        """
+
+        total_lock_unlock_withdraw_amount = f"""
+                        SELECT "holder", SUM(COALESCE(CASE WHEN "event_type" = 0 THEN "amount" ELSE NULL END,
+                                CASE WHEN "event_type" = 1 THEN -amount ELSE NULL END)) AS "lockedAmount",
+                                SUM(CASE WHEN "event_type" = 1 THEN "amount" ELSE 0 END) AS "unlockedAmount",
+                                SUM(CASE WHEN "event_type" = 2 THEN "amount" ELSE 0 END) AS "withdrawnAmount"
+                        FROM (
+                            {union_lock_unlock_withdraw_events}
+                        ) AS "UNION_TABLE"
+                        GROUP BY "holder" ORDER BY "lockedAmount" DESC
+                        """
+
+        leader_board_query = f"""
+                        SELECT ROW_NUMBER() OVER () AS "position",
+                        *
+                        FROM ({total_lock_unlock_withdraw_amount}) AS RESULT_TABLE
+                """
+        return leader_board_query
+
+    @classmethod
+    def get_leader_board(cls, limit: int, offset: int) -> List[Dict]:
+        """
+        Return the leaderboard list ordered by lockedAmount
+
+        :return:
+        """
+        query = f"{cls._get_leader_board_query()} LIMIT {limit} OFFSET {offset}"
+        with connection.cursor() as cursor:
+            cursor.execute(query)
+            return fetch_all_from_cursor(cursor)
+
+    @classmethod
+    def get_holder_position(cls, holder: ChecksumAddress) -> Optional[Dict]:
+        """
+        Get a holder data and position from the leaderboard ordered by lockedAmount
+
+        :return:
+        """
+        query = (
+            f"SELECT * from ({cls._get_leader_board_query()}) AS TEMP WHERE holder=%s"
+        )
+        with connection.cursor() as cursor:
+            holder_address = HexBytes(holder)
+            cursor.execute(query, [holder_address])
+            if result := fetch_all_from_cursor(cursor):
+                return result[0]
+
+    @staticmethod
+    def get_count() -> int:
+        """
+        Return the leaderboard size
+
+        :return:
+        """
+        # LeaderBoard length should be equals than the number of holders of LockEvents
+        return LockEvent.objects.values("holder").distinct().count()
