+++ safe_locking_service/locking_events/views.py
@@ -9,6 +9,7 @@ from gnosis.eth.utils import fast_is_checksum_address
 
 from safe_locking_service import __version__
 from safe_locking_service.locking_events.models import (
+    LeaderBoard,
     LockEvent,
     UnlockEvent,
     WithdrawnEvent,
@@ -88,12 +89,12 @@ class LeaderBoardView(ListAPIView):
     serializer_class = LeaderBoardSerializer
 
     def get_queryset(self, limit, offset):
-        return LockingService.get_leader_board(limit=limit, offset=offset)
+        return LeaderBoard.get_leader_board(limit=limit, offset=offset)
 
     def list(self, request, *args, **kwargs):
         paginator = CustomListPagination(self.request)
         queryset = self.get_queryset(paginator.limit, paginator.offset)
-        paginator.set_count(LockingService.get_leader_board_count())
+        paginator.set_count(LeaderBoard.get_count())
         serializer = LeaderBoardSerializer(queryset, many=True)
 
         return paginator.get_paginated_response(serializer.data)
@@ -110,8 +111,7 @@ class LeaderBoardPositionView(RetrieveAPIView):
     serializer_class = LeaderBoardSerializer
 
     def get_queryset(self, address):
-        locking_service = LockingService(address)
-        return locking_service.get_leader_board_position()
+        return LeaderBoard.get_holder_position(address)
 
     def get(self, request, address, format=None):
         if not fast_is_checksum_address(address):
